+++ tweets/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class TweetConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'tweet'
+    name = 'tweets'
