+++ tweets/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 from . import views
 
-app_name = 'tweet'
+app_name = 'tweets'
 
 urlpatterns = [
   path('create', views.createTweet, name='createTweet'),
